refactor: extract canopy client into bestool-canopy crate
Move the canopy client (mTLS + tailscale probe, status/events POSTs, cert
renewal) out of bestool-alertd and into a new bestool-canopy crate. Both
bestool-alertd and bestool now depend on it directly.
