+++ crates/alertd/Cargo.toml
@@ -18,15 +18,14 @@ required-features = ["cli"]
 
 [dependencies]
 axum = "0.8.9"
+bestool-canopy = { version = "0.1.0", path = "../canopy" }
 bestool-postgres = { version = "1.0.9", path = "../postgres" }
 bytes = "1.9.0"
 clap = { workspace = true, optional = true, features = ["env", "wrap_help"] }
 clap-markdown = { version = "0.1.5", optional = true }
 dirs = "6.0.0"
-flate2 = "1.0.34"
 futures = { workspace = true }
 glob = "0.3.3"
-hickory-resolver = "0.26.1"
 jiff = { version = "0.2.24", features = ["serde"] }
 lloggs = { workspace = true, optional = true }
 mailgun-rs = "2.0.2"
@@ -35,7 +34,6 @@ notify = "8.2.0"
 prometheus = "0.14.0"
 pulldown-cmark = "0.13.3"
 rand.workspace = true
-rcgen = { workspace = true }
 reqwest = { workspace = true }
 serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.145"
@@ -45,7 +43,6 @@ sysinfo.workspace = true
 tempfile = "3.21.0"
 tera = "1.20.0"
 thiserror = { workspace = true }
-time = "0.3"
 tokio = { workspace = true, features = ["full"] }
 tokio-postgres = { version = "0.7.17", features = ["with-jiff-0_2", "with-serde_json-1"] }
 tokio-util = { workspace = true }
