+++ crates/canopy/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "bestool-canopy"
+version = "0.1.0"
+edition = "2024"
+authors = ["Félix Saparelli <felix@passcod.name>", "BES Developers <contact@bes.au>"]
+license = "GPL-3.0-or-later"
+description = "(Internal) BES tooling: Canopy client"
+repository = "https://github.com/beyondessential/bestool/tree/main/crates/canopy"
+exclude.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+flate2 = "1.0.34"
+futures = { workspace = true }
+hickory-resolver = "0.26.1"
+jiff = { version = "0.2.24", features = ["serde"] }
+miette = { workspace = true }
+rcgen = { workspace = true }
+reqwest = { workspace = true }
+serde = { version = "1.0.228", features = ["derive"] }
+serde_json = "1.0.145"
+time = "0.3"
+tokio = { workspace = true, features = ["sync", "time"] }
+tracing = { workspace = true }
